zaps/product/requirements.txt
anyio
asyncpg
backports.zoneinfo
certifi
charset-normalizer
click
databases
dnspython
email-validator
fastapi
#greenlet
h11
idna
psycopg
psycopg2-binary
pyasn1
pydantic
python-dotenv
python-multipart
requests
rsa
six
sniffio
SQLAlchemy
starlette
typing-extensions
urllib3
uvicorn
